Adds link to related paper. (#17)
MSFs as described in "FRP, Refactored" are equivalent to `VarT` up to alpha-renaming. This makes all other properties of MSFs apply to VarTs too.

There is a new paper on MSFs about time transformations and Arrow laws (draft). All of this could be helpful for the user of *varying* too.